Update public API of ajax & refactor package (#1371)
feat(ajax): align / break public API
