+++ packages/ajax/src/interceptors/xsrfHeader.js
@@ -1,4 +1,4 @@
-import './typedef.js';
+import '../typedef.js';
 
 /**
  * @param {string} name the cookie name
@@ -11,21 +11,6 @@ export function getCookie(name, _document = document) {
 }
 
 /**
- * Transforms a request, adding an accept-language header with the current application's locale
- * if it has not already been set.
- * @type {RequestInterceptor}
- */
-export async function acceptLanguageRequestInterceptor(request) {
-  if (!request.headers.has('accept-language')) {
-    const documentLocale = document.documentElement.lang;
-    const localizeLang = document.documentElement.getAttribute('data-localize-lang');
-    const locale = localizeLang || documentLocale || 'en';
-    request.headers.set('accept-language', locale);
-  }
-  return request;
-}
-
-/**
  * Creates a request transformer that adds a XSRF header for protecting
  * against cross-site request forgery.
  * @param {string} cookieName the cookie name
@@ -33,7 +18,7 @@ export async function acceptLanguageRequestInterceptor(request) {
  * @param {Document | { cookie: string }} _document overwriteable for testing
  * @returns {RequestInterceptor}
  */
-export function createXSRFRequestInterceptor(cookieName, headerName, _document = document) {
+export function createXsrfRequestInterceptor(cookieName, headerName, _document = document) {
   /**
    * @type {RequestInterceptor}
    */
