+++ packages/ajax/test/Ajax.test.js
@@ -1,17 +1,17 @@
 import { expect } from '@open-wc/testing';
 import { stub, useFakeTimers } from 'sinon';
-import { AjaxClient, AjaxClientFetchError } from '@lion/ajax';
+import { Ajax, AjaxFetchError } from '@lion/ajax';
 
-describe('AjaxClient', () => {
+describe('Ajax', () => {
   /** @type {import('sinon').SinonStub} */
   let fetchStub;
-  /** @type {AjaxClient} */
+  /** @type {Ajax} */
   let ajax;
 
   beforeEach(() => {
     fetchStub = stub(window, 'fetch');
     fetchStub.returns(Promise.resolve(new Response('mock response')));
-    ajax = new AjaxClient();
+    ajax = new Ajax();
   });
 
   afterEach(() => {
@@ -42,7 +42,7 @@ describe('AjaxClient', () => {
         },
       };
       // When
-      const ajax1 = new AjaxClient(config);
+      const ajax1 = new Ajax(config);
       const result = ajax1.options;
       // Then
       expect(result).to.deep.equal(expected);
@@ -58,7 +58,7 @@ describe('AjaxClient', () => {
       };
       // When
       // @ts-expect-error
-      const ajax1 = new AjaxClient(config);
+      const ajax1 = new Ajax(config);
       const result = ajax1.options?.cacheOptions?.getCacheIdentifier;
       // Then
       expect(result).not.to.be.undefined;
@@ -66,9 +66,9 @@ describe('AjaxClient', () => {
     });
   });
 
-  describe('request()', () => {
+  describe('fetch()', () => {
     it('calls fetch with the given args, returning the result', async () => {
-      const response = await (await ajax.request('/foo', { method: 'POST' })).text();
+      const response = await (await ajax.fetch('/foo', { method: 'POST' })).text();
 
       expect(fetchStub).to.have.been.calledOnce;
       const request = fetchStub.getCall(0).args[0];
@@ -82,9 +82,9 @@ describe('AjaxClient', () => {
 
       let thrown = false;
       try {
-        await ajax.request('/foo');
+        await ajax.fetch('/foo');
       } catch (e) {
-        expect(e).to.be.an.instanceOf(AjaxClientFetchError);
+        expect(e).to.be.an.instanceOf(AjaxFetchError);
         expect(e.request).to.be.an.instanceOf(Request);
         expect(e.response).to.be.an.instanceOf(Response);
         thrown = true;
@@ -97,9 +97,9 @@ describe('AjaxClient', () => {
 
       let thrown = false;
       try {
-        await ajax.request('/foo');
+        await ajax.fetch('/foo');
       } catch (e) {
-        expect(e).to.be.an.instanceOf(AjaxClientFetchError);
+        expect(e).to.be.an.instanceOf(AjaxFetchError);
         expect(e.request).to.be.an.instanceOf(Request);
         expect(e.response).to.be.an.instanceOf(Response);
         thrown = true;
@@ -108,32 +108,32 @@ describe('AjaxClient', () => {
     });
   });
 
-  describe('requestJson', () => {
+  describe('fetchtJson', () => {
     beforeEach(() => {
       fetchStub.returns(Promise.resolve(new Response('{}')));
     });
 
     it('sets json accept header', async () => {
-      await ajax.requestJson('/foo');
+      await ajax.fetchJson('/foo');
       const request = fetchStub.getCall(0).args[0];
       expect(request.headers.get('accept')).to.equal('application/json');
     });
 
     it('decodes response from json', async () => {
       fetchStub.returns(Promise.resolve(new Response('{"a":1,"b":2}')));
-      const response = await ajax.requestJson('/foo');
+      const response = await ajax.fetchJson('/foo');
       expect(response.body).to.eql({ a: 1, b: 2 });
     });
 
     describe('given a request body', () => {
       it('encodes the request body as json', async () => {
-        await ajax.requestJson('/foo', { method: 'POST', body: { a: 1, b: 2 } });
+        await ajax.fetchJson('/foo', { method: 'POST', body: { a: 1, b: 2 } });
         const request = fetchStub.getCall(0).args[0];
         expect(await request.text()).to.equal('{"a":1,"b":2}');
       });
 
       it('sets json content-type header', async () => {
-        await ajax.requestJson('/foo', { method: 'POST', body: { a: 1, b: 2 } });
+        await ajax.fetchJson('/foo', { method: 'POST', body: { a: 1, b: 2 } });
         const request = fetchStub.getCall(0).args[0];
         expect(request.headers.get('content-type')).to.equal('application/json');
       });
@@ -141,9 +141,9 @@ describe('AjaxClient', () => {
 
     describe('given a json prefix', () => {
       it('strips json prefix from response before decoding', async () => {
-        const localAjax = new AjaxClient({ jsonPrefix: '//.,!' });
+        const localAjax = new Ajax({ jsonPrefix: '//.,!' });
         fetchStub.returns(Promise.resolve(new Response('//.,!{"a":1,"b":2}')));
-        const response = await localAjax.requestJson('/foo');
+        const response = await localAjax.fetchJson('/foo');
         expect(response.body).to.eql({ a: 1, b: 2 });
       });
     });
@@ -154,7 +154,7 @@ describe('AjaxClient', () => {
       ajax.addRequestInterceptor(async r => new Request(`${r.url}/intercepted-1`));
       ajax.addRequestInterceptor(async r => new Request(`${r.url}/intercepted-2`));
 
-      await ajax.request('/foo', { method: 'POST' });
+      await ajax.fetch('/foo', { method: 'POST' });
 
       const request = fetchStub.getCall(0).args[0];
       expect(request.url).to.equal(`${window.location.origin}/foo/intercepted-1/intercepted-2`);
@@ -173,7 +173,7 @@ describe('AjaxClient', () => {
         return new Response(`${body} intercepted-2`, { status, statusText, headers });
       });
 
-      const response = await (await ajax.request('/foo', { method: 'POST' })).text();
+      const response = await (await ajax.fetch('/foo', { method: 'POST' })).text();
       expect(response).to.equal('mock response intercepted-1 intercepted-2');
     });
 
@@ -190,7 +190,7 @@ describe('AjaxClient', () => {
       ajax.addRequestInterceptor(interceptor3);
       ajax.removeRequestInterceptor(interceptor1);
 
-      await ajax.request('/foo', { method: 'POST' });
+      await ajax.fetch('/foo', { method: 'POST' });
 
       const request = fetchStub.getCall(0).args[0];
       expect(request.url).to.equal(`${window.location.origin}/foo/intercepted-2/intercepted-3`);
@@ -203,7 +203,7 @@ describe('AjaxClient', () => {
       // @ts-expect-error we're mocking the response as a simple promise which returns a string
       ajax.removeResponseInterceptor(interceptor);
 
-      const response = await ajax.request('/foo', { method: 'POST' });
+      const response = await ajax.fetch('/foo', { method: 'POST' });
       const text = await response.text();
       expect(text).to.equal('mock response');
     });
@@ -211,14 +211,14 @@ describe('AjaxClient', () => {
 
   describe('accept-language header', () => {
     it('is set by default based on localize.locale', async () => {
-      await ajax.request('/foo');
+      await ajax.fetch('/foo');
       const request = fetchStub.getCall(0).args[0];
       expect(request.headers.get('accept-language')).to.equal('en');
     });
 
     it('can be disabled', async () => {
-      const customAjax = new AjaxClient({ addAcceptLanguage: false });
-      await customAjax.request('/foo');
+      const customAjax = new Ajax({ addAcceptLanguage: false });
+      await customAjax.fetch('/foo');
       const request = fetchStub.getCall(0).args[0];
       expect(request.headers.has('accept-language')).to.be.false;
     });
@@ -237,27 +237,27 @@ describe('AjaxClient', () => {
     });
 
     it('XSRF token header is set based on cookie', async () => {
-      await ajax.request('/foo');
+      await ajax.fetch('/foo');
 
       const request = fetchStub.getCall(0).args[0];
       expect(request.headers.get('X-XSRF-TOKEN')).to.equal('1234');
     });
 
     it('XSRF behavior can be disabled', async () => {
-      const customAjax = new AjaxClient({ xsrfCookieName: null, xsrfHeaderName: null });
-      await customAjax.request('/foo');
-      await ajax.request('/foo');
+      const customAjax = new Ajax({ xsrfCookieName: null, xsrfHeaderName: null });
+      await customAjax.fetch('/foo');
+      await ajax.fetch('/foo');
 
       const request = fetchStub.getCall(0).args[0];
       expect(request.headers.has('X-XSRF-TOKEN')).to.be.false;
     });
 
     it('XSRF token header and cookie can be customized', async () => {
-      const customAjax = new AjaxClient({
+      const customAjax = new Ajax({
         xsrfCookieName: 'CSRF-TOKEN',
         xsrfHeaderName: 'X-CSRF-TOKEN',
       });
-      await customAjax.request('/foo');
+      await customAjax.fetch('/foo');
 
       const request = fetchStub.getCall(0).args[0];
       expect(request.headers.get('X-CSRF-TOKEN')).to.equal('5678');
@@ -283,9 +283,9 @@ describe('AjaxClient', () => {
       getCacheIdentifier = () => String(cacheId);
     });
 
-    it('allows configuring cache interceptors on the AjaxClient config', async () => {
+    it('allows configuring cache interceptors on the Ajax config', async () => {
       newCacheId();
-      const customAjax = new AjaxClient({
+      const customAjax = new Ajax({
         cacheOptions: {
           useCache: true,
           timeToLive: 100,
@@ -298,22 +298,22 @@ describe('AjaxClient', () => {
       });
 
       // Smoke test 1: verify caching works
-      await customAjax.request('/foo');
+      await customAjax.fetch('/foo');
       expect(fetchStub.callCount).to.equal(1);
-      await customAjax.request('/foo');
+      await customAjax.fetch('/foo');
       expect(fetchStub.callCount).to.equal(1);
 
       // Smoke test 2: verify caching is invalidated on non-get method
-      await customAjax.request('/foo', { method: 'POST' });
+      await customAjax.fetch('/foo', { method: 'POST' });
       expect(fetchStub.callCount).to.equal(2);
-      await customAjax.request('/foo');
+      await customAjax.fetch('/foo');
       expect(fetchStub.callCount).to.equal(3);
 
       // Smoke test 3: verify caching is invalidated after TTL has passed
-      await customAjax.request('/foo');
+      await customAjax.fetch('/foo');
       expect(fetchStub.callCount).to.equal(3);
       clock.tick(101);
-      await customAjax.request('/foo');
+      await customAjax.fetch('/foo');
       expect(fetchStub.callCount).to.equal(4);
       clock.restore();
     });
@@ -326,7 +326,7 @@ describe('AjaxClient', () => {
       const controller = new AbortController();
       const { signal } = controller;
       // Have to do a "real" request to be able to abort it and verify that this throws
-      const req = ajax.request(new URL('./foo.json', import.meta.url).pathname, {
+      const req = ajax.fetch(new URL('./foo.json', import.meta.url).pathname, {
         method: 'GET',
         signal,
       });
