--- packages/ajax/test/ajax.test.js
@@ -1,14 +1,0 @@
-import { expect } from '@open-wc/testing';
-import { ajax, setAjax, AjaxClient } from '@lion/ajax';
-
-describe('ajax', () => {
-  it('exports an instance of AjaxClient', () => {
-    expect(ajax).to.be.an.instanceOf(AjaxClient);
-  });
-
-  it('can replace ajax with another instance', () => {
-    const newAjax = new AjaxClient();
-    setAjax(newAjax);
-    expect(ajax).to.equal(newAjax);
-  });
-});
